Moved the new files of instruments module into the proper place
