+++ mu4/instruments/instruments.qrc
@@ -10,5 +10,8 @@
         <file>qml/MuseScore/Instruments/StaffSettingsPopup.qml</file>
         <file>qml/MuseScore/Instruments/InstrumentSettingsPopup.qml</file>
         <file>qml/MuseScore/Instruments/InstrumentsDialog.qml</file>
+        <file>qml/MuseScore/Instruments/internal/InstrumentsControlPanel.qml</file>
+        <file>qml/MuseScore/Instruments/internal/InstrumentsTreeItemDelegate.qml</file>
+        <file>qml/MuseScore/Instruments/internal/InstrumentsTreeItemControl.qml</file>
     </qresource>
 </RCC>
