+++ mu4/instruments/instrumentstypes.h
@@ -21,6 +21,7 @@
 
 #include <QVariant>
 #include <vector>
+#include "qobjectdefs.h"
 
 #include "libmscore/interval.h"
 #include "libmscore/drumset.h"
@@ -167,6 +168,24 @@ struct InstrumentsMeta
     InstrumentGroupHash groups;
     InstrumentGenreHash genres;
     MidiArticulationHash articulations;
+};
+
+class InstrumentTreeItemType
+{
+    Q_GADGET
+
+public:
+    enum class ItemType {
+        UNDEFINED = -1,
+        ROOT,
+        PART,
+        INSTRUMENT,
+        STAFF,
+        CONTROL_ADD_STAFF,
+        CONTROL_ADD_DOUBLE_INSTRUMENT
+    };
+
+    Q_ENUM(ItemType)
 };
 }
 }
