+++ mu4/instruments/qml/MuseScore/Instruments/internal/InstrumentsControlPanel.qml
