+++ mu4/instruments/qml/MuseScore/Instruments/internal/InstrumentsTreeItemControl.qml
