+++ mu4/instruments/qml/MuseScore/Instruments/internal/InstrumentsTreeItemDelegate.qml
