+++ mu4/instruments/qml/MuseScore/Instruments/qmldir
@@ -7,3 +7,4 @@ SelectInstrumentDialog 1.0 SelectInstrumentDialog.qml
 InstrumentsPanel 1.0 InstrumentsPanel.qml
 InstrumentSettingsPopup 1.0 InstrumentSettingsPopup.qml
 InstrumentsDialog 1.0 InstrumentsDialog.qml
+InstrumentsPanel 1.0 InstrumentsPanel.qml
